osutil: pass --extrausers option to groupdel
Do not shy off from deleting groups using the `--extrausers` option, if
needed.

See https://bugs.launchpad.net/bugs/1840375

+++ osutil/user.go
@@ -20,6 +20,7 @@
 package osutil
 
 import (
+	"bytes"
 	"fmt"
 	"os"
 	"os/exec"
@@ -150,12 +151,14 @@ func EnsureUserGroup(name string, id uint32, extraUsers bool) error {
 			delCmdStr = append(delCmdStr, "--extrausers")
 		}
 
-		// TODO: groupdel doesn't currently support --extrausers, so
-		// don't try to clean up when it is specified (LP: #1840375)
-		if !extraUsers {
-			delCmdStr = append(delCmdStr, name)
-			cmd = exec.Command(delCmdStr[0], delCmdStr[1:]...)
-			if output2, err2 := cmd.CombinedOutput(); err2 != nil {
+		delCmdStr = append(delCmdStr, name)
+		cmd = exec.Command(delCmdStr[0], delCmdStr[1:]...)
+		if output2, err2 := cmd.CombinedOutput(); err2 != nil {
+			if extraUsers && bytes.Contains(output2, []byte("unrecognized option")) {
+				// We have hit https://bugs.launchpad.net/bugs/1840375
+				// TODO: decide what to do.
+				// The old code was not doing anything here.
+			} else {
 				return fmt.Errorf("groupdel failed with: %s (after %s)", OutputErr(output2, err2), useraddErrStr)
 			}
 		}
